move tests to correct module
