+++ tests/test_clients.py
@@ -1,9 +1,7 @@
-import inspect
 import json
 import sys
 
 import pytest
-from gentools import py2_compatible, return_
 
 import snug
 
@@ -155,128 +153,3 @@ class TestAiohttpSend:
         with mock.patch('aiohttp.client_reqrep.ClientResponse.read', error):
             with pytest.raises(ValueError, match='foo'):
                 loop.run_until_complete(using_aiohttp(req))
-
-
-@py3
-def test__execute_async__(loop):
-    from .py3_only import awaitable
-
-    class StringClient:
-        def __init__(self, mappings):
-            self.mappings = mappings
-
-        def send(self, req):
-            return self.mappings[req]
-
-    snug.send_async.register(StringClient, StringClient.send)
-
-    client = StringClient({
-        'foo/posts/latest': awaitable('redirect:/posts/latest/'),
-        'foo/posts/latest/': awaitable('redirect:/posts/december/'),
-        'foo/posts/december/': awaitable(b'hello world'),
-    })
-
-    class MyQuery:
-
-        @py2_compatible
-        def __iter__(self):
-            redirect = yield '/posts/latest'
-            redirect = yield redirect.split(':')[1]
-            response = yield redirect.split(':')[1]
-            return_(response.decode('ascii'))
-
-    future = snug.Query.__execute_async__(
-        MyQuery(),
-        client=client,
-        authenticate=lambda s: 'foo' + s)
-
-    if sys.version_info > (3, 5):
-        assert inspect.isawaitable(future)
-
-    result = loop.run_until_complete(future)
-    assert result == 'hello world'
-
-
-@py3
-class TestExecuteAsync:
-
-    def test_defaults(self, loop):
-        import asyncio
-        from .py3_only import awaitable
-
-        with mock.patch('snug._async.send_async',
-                        return_value=awaitable(snug.Response(204))) as send:
-
-            @py2_compatible
-            def myquery():
-                return_((yield snug.GET('my/url')))
-
-            future = snug.execute_async(myquery())
-            result = loop.run_until_complete(future)
-            assert result == snug.Response(204)
-            client, req = send.call_args[0]
-            assert isinstance(client, asyncio.AbstractEventLoop)
-            assert req == snug.GET('my/url')
-
-    def test_custom_client(self, loop):
-        from .py3_only import MockAsyncClient
-        client = MockAsyncClient(snug.Response(204))
-
-        @py2_compatible
-        def myquery():
-            return_((yield snug.GET('my/url')))
-
-        future = snug.execute_async(myquery(), client=client)
-        result = loop.run_until_complete(future)
-        assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
-
-    def test_custom_execute(self, loop):
-        from .py3_only import MockAsyncClient
-        client = MockAsyncClient(snug.Response(204))
-
-        class MyQuery:
-            def __execute_async__(self, client, authenticate):
-                return client.send(snug.GET('my/url'))
-
-        future = snug.execute_async(MyQuery(), client=client)
-        result = loop.run_until_complete(future)
-        assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
-
-    def test_auth(self, loop):
-        from .py3_only import MockAsyncClient
-        client = MockAsyncClient(snug.Response(204))
-
-        @py2_compatible
-        def myquery():
-            return_((yield snug.GET('my/url')))
-
-        future = snug.execute_async(myquery(),
-                                    auth=('user', 'pw'),
-                                    client=client)
-        result = loop.run_until_complete(future)
-        assert result == snug.Response(204)
-        assert client.request == snug.GET(
-            'my/url', headers={'Authorization': 'Basic dXNlcjpwdw=='})
-
-    def test_auth_method(self, loop):
-        from .py3_only import MockAsyncClient
-
-        def token_auth(token, request):
-            return request.with_headers({
-                'Authorization': 'Bearer {}'.format(token)
-            })
-
-        @py2_compatible
-        def myquery():
-            return_((yield snug.GET('my/url')))
-
-        client = MockAsyncClient(snug.Response(204))
-        future = snug.execute_async(myquery(), auth='foo', client=client,
-                                    auth_method=token_auth)
-        result = loop.run_until_complete(future)
-
-        assert result == snug.Response(204)
-        assert client.request == snug.GET(
-            'my/url', headers={'Authorization': 'Bearer foo'})
